Split ConstProp pass into own file; propagate lits through nodes
